Move ymd date functions to utils
